requirements.txt
tensorflow==2.17.0
pillow==10.4.0
numpy==1.26.4
flask==3.1.0
opencv-python==4.10.0.82
requests==2.32.3